rename reactReduxConnection to reactRedux
