+++ src/reactRedux/ConnectedComponent.js
